requirements-dev.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
gunicorn>=20.1.0
djangorestframework
django-allauth
django-oauth-toolkit
django-cors-headers
Pillow

# dev_dependencies
django-extensions
pydotplus
drf-spectacular
djangorestframework-simplejwt

#testing
pytest
pytest-django
pytest-cov
factory-boy
